Add function

+++ src/deposit.h
@@ -2,6 +2,6 @@
 #define DEPOSIT_H
 
 float sum_analiz (int days, float sum);
-float correct_analiz (int days, float sum);
+int correct_analiz (int days, float sum);
 
 #endif
